pyproject.toml
[project]
name = "python3-gae-template"
version = "0.1.0"
description = "Template project for python3 on Google App Engine"
dependencies = [
    "fastapi>=0.110.2",
    "mypy>=1.10.0",
    "jinja2>=3.1.3",
    "google-cloud-datastore>=2.19.0",
    "python-multipart>=0.0.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.4.1",
    "pytest>=8.1.1",
    "uvicorn>=0.29.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.rye.scripts]
serve = { cmd = "uvicorn app.main:app --reload --host 0.0.0.0" }
